requirements.txt
boto3==1.4.1
defectdojo_api==0.0.9
PyVirtualDisplay==0.2.1
selenium==2.53.6
python_owasp_zap_v2.4==0.0.8
PyYAML==3.12
